server/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
# TTS Streaming dependencies
anthropic==0.57.1
anyio==4.9.0
attrs==25.3.0
av==14.4.0
certifi
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
coloredlogs==15.0.1
distro==1.9.0
docstring_parser==0.16
eval_type_backport==0.2.2
fastapi==0.115.0
flatbuffers==25.2.10
frozenlist==1.6.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
humanfriendly==10.0
idna==3.10
jiter==0.10.0
livekit>=1.0.19,<2.0.0
# Local forked livekit-agents and plugins
-e ./livekitagents/livekit-agents
-e ./livekitagents/livekit-plugins/livekit-plugins-anthropic
-e ./livekitagents/livekit-plugins/livekit-plugins-elevenlabs
-e ./livekitagents/livekit-plugins/livekit-plugins-openai
-e ./livekitagents/livekit-plugins/livekit-plugins-silero
livekit-api==1.0.7
livekit-protocol>=1.1,<2.0.0
mpmath==1.3.0
multidict==6.4.4
nest-asyncio==1.6.0
numpy==2.2.6
onnxruntime==1.22.1
openai
packaging==25.0
pillow==11.2.1
propcache==0.3.1
protobuf==6.31.1
psutil==7.0.0
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
PyJWT==2.10.1
pyreadline3==3.5.4
python-dotenv==1.1.0
requests==2.32.3
sniffio==1.3.1
sounddevice==0.5.2
starlette==0.38.6
sympy==1.14.0
tqdm==4.67.1
types-protobuf==4.25.0.20240417
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.4.0
uvicorn==0.32.0
watchfiles==1.0.5
websockets==15.0.1
yarl==1.20.0